+++ CHANGELOG.md
@@ -4,5 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0 - 2021-01-05
+### Changed
+- No need to include custom header file, include default XRSound header file.
+- Compiled against XRSound 2.0.
+- Manual
+
 ## 1.0 - 2020-03-01
 - Initial release.